+++ test/test_filters.cpp
@@ -2,7 +2,8 @@
 #include <iostream>
 #include "diplib.h"
 #include "diplib/generation.h"
-#include "diplib/morphology.h"
+//#include "diplib/morphology.h"
+#include "diplib/geometry.h"
 #include "diplib/iterators.h"
 
 template< typename T >
@@ -27,15 +28,18 @@ void PrintPixelValues(
 
 int main() {
    try {
-      dip::Image img{ dip::UnsignedArray{ 20, 15 }, 1, dip::DT_UINT16 };
-      dip::FillDelta( img );
+      dip::Image img{ dip::UnsignedArray{ 20, 15 }, 1, dip::DT_SFLOAT };
+      //dip::FillDelta( img );
+      dip::FillRadiusCoordinate( img );
 
-      PrintPixelValues< dip::uint16 >( img );
+      PrintPixelValues< dip::sfloat >( img );
 
       //dip::Image out = dip::Uniform( img, { { 10, 7 }, "line" } );
-      dip::Image out = dip::Dilation( img, { { 17, 5 }, "diamond" } );
+      //dip::Image out = dip::Dilation( img, { { 17, 5 }, "diamond" } );
+      dip::Image out;
+      dip::Skew( img, out, dip::pi/12.0, 0, 1, "", "" );
 
-      PrintPixelValues< dip::uint16 >( out );
+      PrintPixelValues< dip::sfloat >( out );
 
    } catch( dip::Error e ) {
       std::cout << "DIPlib error: " << e.what() << std::endl;